Move clipboard popup options
