+++ src/clipboard/index.html
